Replaced "unsigned long" type of "seed" argument of function reset() with "unsigned int", because the rest of Random_Generator implementation uses "unsigned int" type.
git-svn-id: svn+ssh://svn.code.sf.net/p/itpp/svn/itpp/trunk@891 f17a810d-92cc-4fca-aeb8-6e32867107aa

+++ itpp/base/random.cpp
@@ -11,7 +11,7 @@
  * IT++ - C++ library of mathematical, signal processing, speech processing,
  *        and communications classes and functions
  *
- * Copyright (C) 1995-2006  (see AUTHORS file for a list of contributors)
+ * Copyright (C) 1995-2007  (see AUTHORS file for a list of contributors)
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -108,7 +108,7 @@ namespace itpp {
 
 
   // Set the seed of the Global Random Number Generator
-  void RNG_reset(unsigned long seed)
+  void RNG_reset(unsigned int seed)
   {
     Random_Generator RNG;
     RNG.reset(seed);
